# API/openapi.yaml
openapi: 3.0.3
info:
  title: UnlikeOtherAuthenticator API
  version: 1.0.0
  description: |
    Centralized OAuth / authentication service. Provides user auth flows,
    token issuance, organisation/team management, and admin operations.

    ## Authentication Schemes

    | Scheme | Header | Format | Used By |
    |--------|--------|--------|---------|
    | Domain Hash Bearer | `Authorization` | `Bearer SHA256(normalizedDomain+clientSecret)` | Domain-scoped server-to-server calls |
    | Admin Bearer | `Authorization` | `Bearer <HS256 JWT>` | Internal admin endpoints |
    | Access Token | `X-UOA-Access-Token` | `Bearer <JWT>` or raw JWT | User-context org endpoints |
    | Config JWT | `X-UOA-Config-JWT` | RS256 JWT | Email send endpoint |

    ## Config URL
    Most endpoints require a `config_url` query parameter pointing to a signed
    RS256 config JWT hosted by the integrator. The server fetches and verifies
    this JWT on every request.
servers:
  - url: https://auth.example.com
    description: Production
security: []
...
components:
  securitySchemes:
    domainHashBearer:
      type: http
      scheme: bearer
      description: |
        `Bearer SHA256(normalizedDomain + clientSecret)` — uses the per-domain client secret, not the shared secret.
    adminBearer:
      type: http
      scheme: bearer
      description: |
        HS256 JWT signed with `ADMIN_ACCESS_TOKEN_SECRET`. Claims must include
        `role: "superuser"` and `domain` matching `ADMIN_AUTH_DOMAIN`.
    accessToken:
      type: apiKey
      in: header
      name: X-UOA-Access-Token
      description: |
        UOA access token (HS256 JWT). May be prefixed with `Bearer ` or sent raw.
    configJwt:
      type: apiKey
      in: header
      name: X-UOA-Config-JWT
      description: RS256 config JWT signed with the domain's private key.
  parameters:
    configUrl:
      name: config_url
      in: query
      required: true
      description: HTTPS URL to a signed RS256 config JWT hosted by the integrator.
      schema:
        type: string
        format: uri
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Generic public error message.
        code:
          type: string
          description: Machine-readable error code (only when DEBUG_ENABLED).
        summary:
          type: string
        details:
          type: array
          items:
            type: string
        hints:
          type: array
          items:
            type: string
...
    TokenResponse:
      type: object
      required: [access_token, expires_in, refresh_token, refresh_token_expires_in, token_type]
      properties:
        access_token:
          type: string
          description: 'HS256 JWT, aud claim is "uoa:access-token".'
        expires_in:
          type: integer
          description: Seconds until access token expiry.
        refresh_token:
          type: string
        refresh_token_expires_in:
          type: integer
          description: Seconds until refresh token expiry.
        token_type:
          type: string
          enum: [Bearer]
        firstLogin:
          type: object
          description: Present only on first-ever login for this user on this domain.
          properties:
            memberships:
              type: object
              properties:
                orgs:
                  type: array
                  items:
                    type: object
                    additionalProperties: true
                teams:
                  type: array
                  items:
                    type: object
                    additionalProperties: true
            pending_invites:
              type: array
              items:
                type: object
                additionalProperties: true
            capabilities:
              type: object
              properties:
                can_create_org:
                  type: boolean
                can_accept_invite:
                  type: boolean
    ConfidentialTokenExchangeResponse:
      type: object
      required: [access_token, issued_token_type, token_type, expires_in, scope]
      additionalProperties: false
      properties:
        access_token:
          type: string
          description: |
            Five-minute RS256 JWT bound to the requested resource. Verify with
            GET /oauth/jwks.json and enforce iss + aud + exp.
        issued_token_type:
          type: string
          enum: ['urn:ietf:params:oauth:token-type:access_token']
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: integer
          enum: [300]
        scope:
          type: string
          enum: [ai.invoke]
...
paths:
...
  /auth/token:
    post:
      tags: [Auth]
      summary: Token exchange (authorization code, refresh token, or confidential JWT assertion)
      operationId: exchangeToken
      description: |
        Exchanges an authorization code or refresh token for the legacy access-token
        pair. A confidential backend may instead use the RFC 8693 token-exchange grant
        with a short-lived source-signed JWT to obtain a five-minute, resource-bound
        RS256 access token. That grant re-resolves the current user, source-domain role,
        and requested ACTIVE org/team membership before issue.
        Requires domain hash bearer authentication.
        Legacy grants are rate limited at 10/min per IP. Confidential exchange is
        isolated from that shared-egress bucket and limited at 600/min per
        authenticated source domain plus 60/min per verified source-domain user.
      security:
        - domainHashBearer: []
      parameters:
        - $ref: '#/components/parameters/configUrl'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - title: Authorization Code Grant
                  type: object
                  required: [code, redirect_url]
                  additionalProperties: false
                  properties:
                    grant_type:
                      type: string
                      enum: [authorization_code]
                      default: authorization_code
                    code:
                      type: string
                      minLength: 1
                    redirect_url:
                      type: string
                      minLength: 1
                    code_verifier:
                      type: string
                      minLength: 1
                      description: PKCE code verifier (required when `code_challenge` was supplied).
                - title: Refresh Token Grant
                  type: object
                  required: [grant_type, refresh_token]
                  additionalProperties: false
                  properties:
                    grant_type:
                      type: string
                      enum: [refresh_token]
                    refresh_token:
                      type: string
                      minLength: 1
                - title: Confidential JWT Token Exchange
                  type: object
                  required: [grant_type, subject_token, subject_token_type, resource]
                  additionalProperties: false
                  properties:
                    grant_type:
                      type: string
                      enum: ['urn:ietf:params:oauth:grant-type:token-exchange']
                    subject_token:
                      type: string
                      minLength: 1
                      maxLength: 16384
                      description: |
                        RS256 JWT signed by a kid in the source config's same-host
                        jwks_url. Required claims: iss and source_domain equal the
                        source config domain; aud equals this /auth/token endpoint;
                        stable UOA sub; active { orgId, teamId }; jti; iat; exp.
                        exp-iat must not exceed 60 seconds.
                    subject_token_type:
                      type: string
                      enum: ['urn:ietf:params:oauth:token-type:jwt']
                    resource:
                      type: string
                      maxLength: 2048
                      description: |
                        Exact configured source-domain-to-resource allowlist target.
                        This value becomes the issued access token's aud.
      responses:
        '200':
          description: Legacy token pair or confidential resource access token issued.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/TokenResponse'
                  - $ref: '#/components/schemas/ConfidentialTokenExchangeResponse'
        '400':
          description: Invalid request or token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid domain hash auth.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limited.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...